Lua exceptions are now be prettified during mod load.
There is now the ScriptException, which can be thrown by the
ScriptEnvironment. The AbstractThreadedUpdatable is now also handling
exceptions which are thrown during its init method correctly, and does
not start (instead of staying in an undefined state).

Last but not least, the server does now exit when the server environment
did fail to start.